+++ examples/start_gui.py
@@ -4,7 +4,9 @@ script to start the GUI
 # pragma: no cover
 from __future__ import annotations
 
+import logging
 import sys
+from functools import partial
 from pathlib import Path
 from platform import system
 from sys import argv
@@ -107,6 +109,7 @@ class GUI(GuiStructure):
             maximal_value=200,
             category=self.category_inputs
         )
+        self.int_a.change_event(self.disable_aim(self.aim_sub, self.page_inputs, partial(self.int_a.check_linked_value, (None, 5))))
 
         self.int_units = els.IntBoxWithUnits(
             label="IntBoxWithUnits",
@@ -139,13 +142,25 @@ class GUI(GuiStructure):
             category=self.sub_category,
         )
 
-        self.list_box = els.ListBox(label="List box", default_index=0, category=self.category_inputs, entries=["1","2","3","4"])
+        self.list_box = els.ListBox(label="List box", default_index=0, category=self.category_inputs, entries=["1", "2", "3", "4"])
         folder: Path = Path(__file__).parent
         file = f'{folder.joinpath("./example_data.csv")}'
         self.filename = els.FileNameBox(label="Filename", default_value=file, category=self.category_inputs, dialog_text="Hello", error_text="no file found",
                                         file_extension=["txt", "csv"])
 
-        self.button_box = els.ButtonBox(label="a or b?", default_index=0, entries=["a", "b"], category=self.category_inputs)
+        self.text_box_only_on_add = els.TextBox(label="Only visible on add", default_text="Hello", category=self.category_inputs)
+
+        self.aim_add.add_link_2_show(self.text_box_only_on_add)
+        # self.aim_add.add_link_2_show(self.filename)
+        self.button_box = els.ButtonBox(label="a or b or c?", default_index=0, entries=["a", "b", "c"], category=self.category_inputs)
+
+        self.aim_plot.widget.toggled.connect(self.disable_button_box(self.button_box, at_index=2, func_2_check=self.aim_plot.widget.isChecked))
+        self.float_b.change_event(self.disable_button_box(self.button_box, 1, partial(self.float_b.check_linked_value, (50, None))))
+        self.int_a.change_event(self.disable_button_box(self.button_box, 0, partial(self.int_a.check_linked_value, (None, 10))))
+
+        self.button_box_short = els.ButtonBox(label="b or c?", default_index=0, entries=["b", "c"], category=self.category_inputs)
+        self.float_b.change_event(self.disable_button_box(self.button_box_short, 1, partial(self.float_b.check_linked_value, (50, None))))
+        self.int_a.change_event(self.disable_button_box(self.button_box_short, 0, partial(self.int_a.check_linked_value, (None, 10))))
 
         self.function_button = els.FunctionButton(button_text="function", icon="Add", category=self.category_inputs)
 
@@ -153,16 +168,20 @@ class GUI(GuiStructure):
         self.text_box_multi_line = els.TextBoxMultiLine(label="Example Multi Line", default_text="Hello\nmulti line", category=self.category_inputs)
         self.text_box.deactivate_size_limit()
         self.pass_word = els.TextBox(label="Password", default_text="1234", category=self.category_inputs, password=True)
-        
+
         self.flex_option = els.FlexibleAmount(label=self.translations.flex_option, default_length=2, entry_mame="Layer", category=self.category_inputs,
-                                              min_length=2,default_values=[["layer 1", 9.5, 3, 2], ["layer 2", 10.5, 2, 1]])
+                                              min_length=2, default_values=[["layer 1", 9.5, 3, 2], ["layer 2", 10.5, 2, 1]])
         self.flex_option.add_option(els.TextBox, name="name", default_text="layer")
-        self.flex_option.add_option(els.FloatBox, name="thickness", default_value=10, minimal_value=5, decimal_number = 2)
+
+        self.flex_option.add_option(els.FloatBox, name="thickness", default_value=10, minimal_value=5, decimal_number=2)
         self.flex_option.add_option(els.IntBox, name="amount", default_value=4, minimal_value=2)
         self.flex_option.add_option(els.ListBox, name="entry", default_index=0, entries=["entry 1", "entry 2", "entry 3"])
         self.hint_flex = els.Hint(hint="wrong length of flexible option", category=self.category_inputs, warning=True)
         self.flex_option.add_link_2_show(self.hint_flex, 2, 6)
         self.aim_plot.add_link_2_show(self.flex_option)
+        # self.button_box.add_link_2_show(self.filename, on_index=1)
+        self.show_option_under_multiple_conditions(self.filename, [self.button_box, self.aim_add], [partial(self.button_box.check_linked_value, 1),
+                                                                                                    self.aim_add.widget.isChecked])
 
         self.category_grid = els.Category(page=self.page_inputs, label="Grid")
         self.category_grid.activate_grid_layout(3)
@@ -216,7 +235,7 @@ class GUI(GuiStructure):
         self.result_text_sub.function_to_convert_to_text(lambda x: round(x, 2))
 
         self.result_export = els.ResultExport("Export results", icon="Download", category=self.numerical_results, export_function=ResultsClass.export,
-                                          caption="Select file", file_extension="txt")
+                                              caption="Select file", file_extension="txt")
 
         self.figure_results = els.ResultFigure(label=self.translations.figure_results, page=self.page_result, x_axes_text="X-Axes", y_axes_text="Y-Axes")
         self.legend_figure_results = els.FigureOption(
@@ -225,13 +244,13 @@ class GUI(GuiStructure):
 
         self.figure_results.fig_to_be_shown(class_name="ResultsClass", function_name="create_plot")
 
-        self.figure_results_multiple_lines = els.ResultFigure(label=self.translations.figure_results_multiple_lines, page=self.page_result, x_axes_text="X-Axes", y_axes_text="Y-Axes")
+        self.figure_results_multiple_lines = els.ResultFigure(label=self.translations.figure_results_multiple_lines, page=self.page_result,
+                                                              x_axes_text="X-Axes", y_axes_text="Y-Axes")
         self.legend_figure_results_multiple_lines = els.FigureOption(
             category=self.figure_results_multiple_lines, label="Legend on", param="legend", default=0, entries=["No", "Yes"],
             entries_values=[False, True]
         )
         self.figure_results_multiple_lines.fig_to_be_shown(class_name="ResultsClass", function_name="create_plot_multiple_lines")
-
 
         self.aim_add.add_link_2_show(self.result_text_add)
         self.aim_add.add_link_2_show(self.result_export)
@@ -244,6 +263,10 @@ class GUI(GuiStructure):
         self.page_result.set_previous_page(self.page_inputs)
         self.page_result.set_next_page(self.page_settings)
         self.page_settings.set_previous_page(self.page_result)
+
+    def check(self) -> bool:
+        if self.started:
+            logging.info('This should not be shown whilst loading')
 
 
 def run(path_list=None):  # pragma: no cover